% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util___addLags.R
\name{addLags}
\alias{addLags}
\title{Add lagged versions to all numeric columns in a data frame}
\usage{
addLags(df, lag, .zero = TRUE)
}
\arguments{
\item{df}{(tibble) Data frame with data to lag}

\item{lag}{(integer) Maximum number of lags to return}

\item{.zero}{(boolean) Return original non-lagged data-frame? TRUE by default.}
}
\description{
Add lagged versions to all numeric columns in a data frame
}
