% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util__transforms.R
\name{diff}
\alias{diff}
\title{Conducts common time series transformations - differencing, log-differencing, moving-averages, and lagged moving averages}
\usage{
diff(x, .l)
}
\arguments{
\item{x}{(numeric vector) A vector of data to be transformed}

\item{.l}{(integer) Number of laggs for taking a difference}

\item{.length}{(integer) Length of lags for taking moving average}
}
\description{
Conducts common time series transformations - differencing, log-differencing, moving-averages, and lagged moving averages
}
