% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping__getDataFred.R
\name{getDataFred}
\alias{getDataFred}
\title{Scrape data from FRED.}
\usage{
getDataFred(seriesId, apiKey, .freq = NULL, .returnVintages = FALSE)
}
\arguments{
\item{seriesId}{(string) FRED ID of economic variable}

\item{apiKey}{(string) User's FRED API Key}

\item{.freq}{(string) One of 'd', 'm', 'q'. If NULL, returns highest available frequency.}

\item{.returnVintages}{(boolean) If TRUE, returns all historic forecast values ('vintages').}
}
\value{
A data frame of forecasts
}
\description{
Scrape data from FRED.
}
